Format

Customizable generator that you can specify your own bytes-array-creation method. You can specify any R function that recieves `size` as its argument and returns a bytes-array of wihch length is `size`.

Arguments

size
length of output.

dict
choose one of c("url", "numbers", "lowercase", "uppercase", "nolookalikes") or give acceptable strings set as character.

init.locales
vocabulary sets of words. defaut sets are c("en", "ja").

use_func
function name that will be used for generating buffer array.

Value

strings output.