NanoID class

R6 class definition of NanoID-class.

Public fields

ctx

V8 context.

Methods

Public methods

NanoIdGood$new()

NanoIdGood$format()

NanoIdGood$generate()

NanoIdGood$nonsecure()

NanoIdGood$clone()

Method new()

Constructor

Usage

NanoIdGood$new(ctx)

Arguments

ctx

V8 context.

Method format()

Format

Usage

NanoIdGood$format(
  size = 21L,
  dict = c("url", "numbers", "lowercase", "uppercase", "nolookalikes"),
  init.locales = c("en", "ja"),
  use_func = "randombytes"
)

Arguments

size

length of output.

dict

choose one of c("url", "numbers", "lowercase", "uppercase", "nolookalikes") or give acceptable strings set as character.

init.locales

vocabulary sets of words. default sets are c("en", "ja").

use_func

function name that will be used for generating bytes array.

Details

Customizable generator that you can specify your own bytes-array-creation method. You can specify any R function that receives `size` as its argument and returns a bytes-array of which length is `size`.

Returns

strings output.

Method generate()

Generate

Usage

NanoIdGood$generate(
  size = 21L,
  dict = c("url", "numbers", "lowercase", "uppercase", "nolookalikes"),
  init.locales = c("en", "ja")
)

Arguments

size

length of output.

dict

choose one of c("url", "numbers", "lowercase", "uppercase", "nolookalikes") or give acceptable strings set as character.

init.locales

vocabulary sets of obscene words. default sets are c("en", "ja").

Returns

strings output.

Method nonsecure()

Nonsecure

Usage

NanoIdGood$nonsecure(size = 21L, init.locales = c("en", "ja"))

Arguments

size

length of output.

init.locales

vocabulary sets of obscene words. default sets are c("en", "ja").

Details

Faster but non-secure version of NanoID generator (bigger collision probability).

Returns

strings output.

Method clone()

The objects of this class are cloneable with this method.

Usage

NanoIdGood$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.